requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
stripe==5.5.0
APScheduler==3.10.4
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
anthropic==0.25.1
python-jose[cryptography]==3.3.0
itsdangerous==2.1.2
pydantic-settings>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0